use mode instead configuration, implement in build-ios
